move User.canTakeModules to a repo function
