+++ packages/database/src/entity/User.ts
@@ -38,39 +38,4 @@ export class User {
 
   @Column()
   graduationSemester: number
-
-  /**
-   * Given a module code, checks if user has cleared sufficient pre-requisites.
-   * Currently does not check for preclusion.
-   *
-   * @param{string} moduleCode
-   * @return{Promise<boolean>}
-   */
-  async canTakeModule(moduleCode: string): Promise<boolean | void> {
-    return await container(async () => {
-      // find module
-      const module = await ModuleRepository.findOne({
-        where: {
-          moduleCode: moduleCode,
-        },
-      })
-
-      // Relations are not stored in the entity, so they must be explicitly
-      // asked for from the DB
-      const user = await UserRepository.findOne({
-        where: {
-          id: this.id,
-        },
-        relations: ['modulesDone'],
-      })
-      const modulesDone = user.modulesDone
-
-      // check if PrereqTree is fulfilled
-      const completedModulesCodes = modulesDone.map(
-        (one: Module) => one.moduleCode
-      )
-
-      return utils.checkTree(module.prereqTree, completedModulesCodes)
-    })
-  }
 }
