+++ packages/database/tests/user.test.ts
@@ -24,7 +24,9 @@ test('canTakeModule is successful', async () => {
         },
       })
       const modulesTested = ['MA2101', 'MA1100', 'CS2040S', 'CS1010S']
-      return Promise.all(modulesTested.map((x) => user.canTakeModule(x)))
+      return Promise.all(
+        modulesTested.map((x) => UserRepository.canTakeModule(user, x))
+      )
     })
   )
   expect(res).toStrictEqual([true, false, false, true])
